+++ Sources/Kore/IO/FileReader.h
@@ -14,11 +14,11 @@ namespace Kore {
 		~FileReader();
 		bool open(const char *filename, FileType type = Asset);
 		void close();
-		int read(void *data, int size) override;
+		size_t read(void *data, size_t size) override;
 		void *readAll() override;
-		int size() override;
-		int pos() override;
-		void seek(int pos) override;
+		size_t size() override;
+		size_t pos() override;
+		void seek(size_t pos) override;
 
 		kinc_file_reader_t reader;
 		FileType type;
